feat(api): new endpoint that returns the languages used at one location
